v3.5: Update everything to the last archived copy I had.
This is probably the version Jeff worked on upon returning to working
on DeckMaker briefly, around 1998 or so? This version returns back to
being written in C rather than C++ (I think it wasn't a terribly difficult
transition, with the C++ being rather meek).

This version *might* build? I do not currently have the tools to verify.