v3.2 Brief detour into C++ version.
This is allegedly version 3.2, according to the main .cpp file. This would
be the first version not completely written by Jeff. These files were just
mingled in with the later C versions in my local copy, there was no version
control for any of this at the time.

I am unsure if this is a real released version or just a detour into a C++
rewrite that didn't pan out. This is likely incomplete, as later changes
to the header files would have broken it and there's no build files from
this period.

I believe, though, that I may have been working on it in C++ until Jeff Heskett
returned to it later on, for the 3.5 release, which is the version I have complete
files for.